complete user society flow, hibernate optimization and stuff

+++ src/main/java/com/agrocom/dao/RoleDAO.java
@@ -4,5 +4,5 @@ import com.agrocom.model.Role;
 
 public interface RoleDAO {
 
-    public Role getRole(int id);
+    public Role getRole(Long id);
 }
