+++ src/main/java/com/agrocom/dao/UserDAO.java
@@ -18,7 +18,7 @@ public interface UserDAO {
 
     List searchUserByLastName(String lastName);
 
-    Integer addUser(User user);
+    Long addUser(User user);
 
     Boolean updateUser(User user);
