+++ src/main/java/com/agrocom/service/UserService.java
@@ -20,7 +20,7 @@ public interface UserService {
 
     List searchUserByLastName(String lastName);
 
-    Integer addUser(User user);
+    Long addUser(User user);
 
     Boolean updateUser(User user);
